+++ src/components/AboutCard.vue
@@ -5,7 +5,7 @@ defineProps<{
 </script>
 
 <template>
-  <div class="p-6 bg-gray-50 dark:bg-gray-950 rounded-lg">
+  <div class="p-6 bg-gray-50 dark:bg-gray-950 rounded-lg shadow-lg">
     <h3 class="text-xl font-semibold mb-4">
       {{ title }}
     </h3>
