+++ src/components/ContactCard.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="p-6 bg-gray-50 dark:bg-gray-950 rounded-lg">
+  <div class="p-6 bg-gray-50 dark:bg-gray-950 rounded-lg shadow-lg">
     <div class="flex flex-col justify-center gap-1.5 mx-auto">
       <slot />
     </div>
