+++ src/components/Logo/DevMadeItLogo.vue
@@ -5,7 +5,7 @@
         DEV
       </span> Made IT
     </div>
-    <div class="text-transparent bg-clip-text bg-gradient-to-b from-gray-500 to-gray-200 opacity-80">
+    <div class="text-gray-500 dark:text-gray-400 opacity-80">
       Madeyski Jan
     </div>
   </div>
