+++ src/components/Page/About.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { useI18n } from 'vue-i18n'
-import AboutCard from './AboutCard.vue'
+import AboutCard from '../AboutCard.vue'
 
 const { t } = useI18n()
 </script>
@@ -8,7 +8,7 @@ const { t } = useI18n()
 <template>
   <section id="about" class="py-20 bg-white dark:bg-gray-900">
     <div class="container mx-auto px-4">
-      <h2 class="text-4xl font-bold text-center mb-12">
+      <h2 class="text-4xl font-bold text-center mb-12 border-b-4 border-primary pb-2 w-fit mx-auto">
         {{ t('about.title') }}
       </h2>
       <div class="grid md:grid-cols-3 gap-8">
