+++ src/components/Page/Contact.vue
@@ -2,7 +2,7 @@
 import { Briefcase, Building, Github, Linkedin, Mail, MapPin, Phone, ShieldCheck } from 'lucide-vue-next'
 import { useI18n } from 'vue-i18n'
 import { config } from '@/config'
-import CardComponent from '../CardComponent.vue'
+import CardComponent from '../ContactCard.vue'
 
 const { t } = useI18n()
 </script>
@@ -10,7 +10,7 @@ const { t } = useI18n()
 <template>
   <section id="contact" class="py-20 bg-white dark:bg-gray-900">
     <div class="container mx-auto px-4">
-      <h2 class="text-4xl font-bold text-center mb-12">
+      <h2 class="text-4xl font-bold text-center mb-12 border-b-4 border-primary pb-2 w-fit mx-auto">
         {{ t('contact.title') }}
       </h2>
       <div class="grid md:grid-cols-3 gap-8">
